moved the paging function to a separate js file
