+++ app/javascript/packs/materials/index.js
@@ -2,4 +2,5 @@ import 'bootstrap/js/dist/modal'
 import 'bootstrap/js/dist/tooltip'
 import '../../application/tooltip_initializer'
 import '../../application/locale_config'
+import '../../application/paging_handler'
 import '../../materials/index'
